[go] Move type helper functions to types module
This allows us to simplify the tests again because these functions
don't rely on the config module.
