+++ src/go/types.ts
@@ -108,3 +108,78 @@ export function filterMapBoard<T>(
     }
     return result;
 }
+
+/**
+ * Type predicate for validating a string is a Vertex.
+ *
+ * @param s - candidate Vertex string
+ * @returns Whether the candidate is a valid Vertex
+ */
+export function isVertex(s: string): s is Vertex {
+    if (s === 'pass') return true;
+    const x = columnIndex(s);
+    const y = rowIndex(s);
+    return x !== -1 && y !== -1;
+}
+
+/**
+ * Convert a GTP vertex to a 0-based multidimensional board array index.
+ *
+ * @param vertex - Vertex to convert
+ * @returns x and y indices
+ */
+export function toIndices(vertex: Vertex): [number, number] {
+    if (vertex === 'pass') return [-1, -1];
+
+    const x = columnIndex(vertex);
+    if (x === -1)
+        throw new Error(`tried to transform invalid vertex ${vertex}`);
+
+    const y = rowIndex(vertex);
+    if (y === -1)
+        throw new Error(`tried to transform invalid vertex ${vertex}`);
+
+    return [x, y];
+}
+
+/**
+ * Convert a 0-based multidimensional board array index to a GTP vertex.
+ *
+ * @param x - Column index
+ * @param y - Row index
+ * @returns Vertex corresponding to indices
+ */
+export function toVertex(x: number, y: number): Vertex {
+    if (x >= COL_NAMES.length)
+        throw new Error(`tried to generate vertex with invalid col index ${x}`);
+
+    if (y >= ROW_NAMES.length)
+        throw new Error(`tried to generate vertex with invalid row index ${y}`);
+
+    const col = COL_NAMES[x];
+    const row = ROW_NAMES[y];
+    return `${col}${row}` as Vertex;
+}
+
+/**
+ * Translate a Vertex string to the corresponding column index.
+ *
+ * @param s - Vertex string
+ * @returns zero based column index
+ */
+export function columnIndex(s: string): number {
+    return COL_NAMES.findIndex((colName) => s.startsWith(colName));
+}
+
+/**
+ * Translate a Vertex string to the correpsonding row index.
+ *
+ * @param s - Vertex string
+ * @returns zero based row index
+ */
+export function rowIndex(s: string): number {
+    const match = s.match(/(\d+)$/);
+    if (!match) return -1;
+    const row = Number.parseInt(match[1], 10);
+    return ROW_NAMES.findIndex((rowName) => rowName === row);
+}
